requirements.txt
llama-cpp-python
torch --index-url https://download.pytorch.org/whl/cu121
torchvision
torchaudio
transformers
accelerate
huggingface_hub
beautifulsoup4
googlesearch-python
pandas
tqdm
trafilatura
requests
lxml